+++ src/main/kotlin/net/alexanderkahn/longball/service/model/LineupPosition.kt
@@ -1,4 +1,4 @@
 package net.alexanderkahn.longball.service.model
 
-data class LineupPosition(val player: Long, val battingOrder: Long, val fieldPosition: FieldPosition)
+data class LineupPosition(val player: Long, val battingOrder: Short, val fieldPosition: FieldPosition)
